add authorization for all ExpenseController methods

+++ src/main/java/com/myWallet/controllers/ExpenseController.java
@@ -2,9 +2,12 @@ package com.myWallet.controllers;
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -13,7 +16,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.myWallet.dto.ExpenseDto;
+import com.myWallet.model.AppUser;
 import com.myWallet.services.ExpenseService;
+import com.myWallet.services.TokenService;
 
 @RequestMapping(value = "/api/expense")
 @RestController
@@ -22,33 +27,82 @@ public class ExpenseController {
 	@Autowired
 	private ExpenseService expenseService;
 	
+	@Autowired
+	private TokenService tokenService;
+	
 	@RequestMapping(value = "/list", method = RequestMethod.GET)
-	public List<ExpenseDto> getExpenseList(@RequestParam(value = "categoryName", required = false) String categoryName) {
-		if (categoryName==null) {
-			return expenseService.getListExpense();
-		}
-		else {
-			return expenseService.getListExpenseByCategoryName(categoryName);
+	public List<ExpenseDto> getExpenseList(@RequestParam(value = "categoryName", required = false) String categoryName, HttpServletRequest request, HttpServletResponse response) {
+		String token = request.getHeader("Authorization");
+		AppUser user = tokenService.validateToken(token);
+		if (user == null) {
+			response.setStatus(HttpStatus.UNAUTHORIZED.value());
+			return null;
+		} else {
+			response.setStatus(HttpStatus.OK.value());
+			return expenseService.getListExpense(user);
 		}
 	}
 	
-	@RequestMapping(value = "/list/{sortType}", method = RequestMethod.GET)
-	public List<ExpenseDto> getSortedExpenseList(@PathVariable String sortType) {
-		return expenseService.getSortedListExpense(sortType);
+	@RequestMapping(value = "/list/category/{categoryName}", method = RequestMethod.GET)
+	public List<ExpenseDto> getExpenseListByCategoryName(@PathVariable String categoryName, HttpServletRequest request, HttpServletResponse response) {
+		String token = request.getHeader("Authorization");
+		AppUser user = tokenService.validateToken(token);
+		if (user == null) {
+			response.setStatus(HttpStatus.UNAUTHORIZED.value());
+			return null;
+		} else {
+			response.setStatus(HttpStatus.OK.value());
+			return expenseService.getListExpenseByCategoryName(user, categoryName);
+		}
+	}
+	
+	@RequestMapping(value = "/list/sort/{sortType}", method = RequestMethod.GET)
+	public List<ExpenseDto> getSortedExpenseList(@PathVariable String sortType, HttpServletRequest request, HttpServletResponse response) {
+		String token = request.getHeader("Authorization");
+		AppUser user = tokenService.validateToken(token);
+		if (user == null) {
+			response.setStatus(HttpStatus.UNAUTHORIZED.value());
+			return null;
+		} else {
+			response.setStatus(HttpStatus.OK.value());
+			return expenseService.getSortedListExpense(user, sortType);
+		}
 	}
 	
 	@RequestMapping(method = RequestMethod.POST)
-	public void addExpense(@RequestBody @Valid ExpenseDto expenseDto) {
-		expenseService.addExpense(expenseDto);
+	public void addExpense(@RequestBody @Valid ExpenseDto expenseDto, HttpServletRequest request, HttpServletResponse response) {
+		String token = request.getHeader("Authorization");
+		AppUser user = tokenService.validateToken(token);
+		if (user == null) {
+			response.setStatus(HttpStatus.UNAUTHORIZED.value());
+		} else {
+			response.setStatus(HttpStatus.CREATED.value());
+			expenseService.addExpense(expenseDto);
+		}
 	}
 	
 	@RequestMapping(value = "/{expenseID}", method = RequestMethod.DELETE)
-	public void deleteExpense(@PathVariable Long expenseID) {
-		expenseService.deleteExpense(expenseID);
+	public void deleteExpense(@PathVariable Long expenseID, HttpServletRequest request, HttpServletResponse response) {
+		String token = request.getHeader("Authorization");
+		AppUser user = tokenService.validateToken(token);
+		if (user == null) {
+			response.setStatus(HttpStatus.UNAUTHORIZED.value());
+		} else {
+			response.setStatus(HttpStatus.OK.value());
+			expenseService.deleteExpense(expenseID);
+		}
 	}
 	
 	@RequestMapping(value = "/{expenseID}", method = RequestMethod.GET)
-	public ExpenseDto getExpense(@PathVariable Long expenseID) {
-		return expenseService.getExpense(expenseID);
+	public ExpenseDto getExpense(@PathVariable Long expenseID, HttpServletRequest request, HttpServletResponse response) {
+		String token = request.getHeader("Authorization");
+		AppUser user = tokenService.validateToken(token);
+		if (user == null) {
+			response.setStatus(HttpStatus.UNAUTHORIZED.value());
+			return null;
+		} else {
+			response.setStatus(HttpStatus.OK.value());
+			return expenseService.getExpense(expenseID);
+		}
 	}
 }
